Updates
Fixed class. Also fixed the command details for frisking. I also added a
way so the command can not be run in the console to prevent server
crashing; /frisk HAS to be run in-game only. Now I'm going to implement
this class for the command instead of having it all in the main class,
making it more OOP. Oh, and i spaced it out a bit.